requirements.txt
elasticsearch==8.14.0
kafka_python==2.0.2
